courses endpoint auth

+++ courseswebsite/src/main/java/com/medhat/springboot/courseswebsite/rest/CoursesRestController.java
@@ -2,10 +2,16 @@ package com.medhat.springboot.courseswebsite.rest;
 
 
 import com.medhat.springboot.courseswebsite.entity.Course;
+import com.medhat.springboot.courseswebsite.exception.NotAuthorizedException;
+import com.medhat.springboot.courseswebsite.securingweb.WebSecurityPermissions;
 import com.medhat.springboot.courseswebsite.service.CoursesService;
+import com.medhat.springboot.courseswebsite.service.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @RestController
@@ -14,9 +20,12 @@ public class CoursesRestController {
 
     private CoursesService coursesService;
 
+    private UsersService usersService;
+
     @Autowired
-    public CoursesRestController(CoursesService coursesService) {
+    public CoursesRestController(CoursesService coursesService,UsersService usersService) {
         this.coursesService = coursesService;
+        this.usersService = usersService;
     }
 
     @GetMapping("/courses")
@@ -31,20 +40,37 @@ public class CoursesRestController {
 
     @PostMapping("/courses")
     public Course addCourse(@RequestBody Course course){
-        course.setId(0);
-        coursesService.saveCourse(course);
+        if(WebSecurityPermissions.hasRole("INSTRUCTOR")){
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String currentPrincipalName = authentication.getName();
+            course.setInstructorId(usersService.getByUserName(currentPrincipalName).getId());
+            coursesService.saveCourse(course);
+        }
+        else{
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
+        }
         return course;
     }
 
-    @PutMapping("/courses")
-    public Course updateCourse(@RequestBody Course course){
-        coursesService.saveCourse(course);
-        return course;
-    }
+//    @PutMapping("/courses")
+//    public Course updateCourse(@RequestBody Course course){
+//        if(WebSecurityPermissions.hasRole("INSTRUCTOR")){
+//            coursesService.saveCourse(course);
+//        }
+//        else{
+//            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
+//        }
+//        return course;
+//    }
 
     @DeleteMapping("/courses/{courseId}")
     public int deleteById(@PathVariable int courseId){
-        coursesService.deleteById(courseId);
+        if(WebSecurityPermissions.hasRole("ADMIN")){
+            coursesService.deleteById(courseId);
+        }
+        else{
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
+        }
         return courseId;
     }
